third_party/blink/web_tests/fast/css/forced-colors-mode/forced-colors-mode-09.html

<!DOCTYPE html>
<html>
<head>
<title>
  Forced colors mode - forced-colors media query.
  Tests that styles defined inside or outside a forced-colors media query are
  overridden in forced colors mode.
</title>
<style type="text/css">
  @media (forced-colors) {
    p {
      color: rgb(0, 0, 255);
    }
  }
  p {
    color: rgb(0, 0, 255);
  }
</style>
</head>
<body>
  <p id="p">
    This text should NOT be blue forced colors mode.
  </p>
</body>

<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script>
  test(function() {
    assert_not_equals(getComputedStyle(p).color, "rgb(0, 0, 255)");
  }, "Checks that styles defined inside/outside forced-colors are overridden in forced colors mode.");
  </script>
</html>
